RVB-002: Separate app into modules
